商品品牌增删改查和商品类型左端树形菜单

+++ gxshopping-basic/basic-util/src/main/java/yahaha/gxshopping/util/PageList.java
@@ -5,23 +5,23 @@ import java.util.List;
 
 public class PageList<T> {
 
-    private Integer total;
+    private Long total;
 
     private List<T> rows = new ArrayList<>();
 
     public PageList() {
     }
 
-    public PageList(Integer total, List<T> rows) {
+    public PageList(Long total, List<T> rows) {
         this.total = total;
         this.rows = rows;
     }
 
-    public Integer getTotal() {
+    public Long getTotal() {
         return total;
     }
 
-    public void setTotal(Integer total) {
+    public void setTotal(Long total) {
         this.total = total;
     }
